requirements.txt
langchain-nvidia-ai-endpoints==0.3.10
langchain
fastapi==0.115.12
python-dotenv==1.1.0
uvicorn[standard]
PyYAML==6.0.2
langchain-tavily
pygithub
python-gitlab
langchain-google-community[gmail]
O365
beautifulsoup4
twilio
playwright
lxml
SQLAlchemy
pyspark
ads4gpts-langchain
langchain-agentql
langchain-apify
arxiv
langchain-discord
elevenlabs
# langchain-financial-datasets
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
langchain-googledrive
google-search-results
langchain-mcp-adapters
langgraph
langsmith>=0.3.42
langgraph>=0.4.2
langgraph_sdk>=0.1.66
langgraph-checkpoint>=2.0.25
langgraph-prebuilt>=0.1.8
langchain-mcp-adapters>=0.0.10

# Tools & Specific Integrations
langchain-googledrive>=0.3.35
langchain-google-community>=2.0.7
langchain-tavily>=0.1.6
langchain-discord>=1.0.2
langchain-apify>=0.1.2
serpapi>=0.1.1
google-search-results>=2.4.2
arxiv>=2.2.0
feedparser>=6.0.11
beautifulsoup4>=4.13.4
lxml>=5.4.0
tiktoken>=0.9.0
playwright>=1.52.0
playwright-stealth>=1.0.6
agentql>=1.9.2
numexpr 
faiss-cpu 